adds socket into hook and new page for canvas
Fixes the guest bug 🥳
